% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{apply_restrictions}
\alias{apply_restrictions}
\title{Filter a grid based on a list of restrictions}
\usage{
apply_restrictions(grid, comp_restrict)
}
\arguments{
\item{grid}{A \code{data.frame} filled with TRUE and FALSEs.}

\item{comp_restrict}{A list with elements:}
}
\value{
The input grid filtered according to the restrictions passed in \code{comp_restrict}.
}
\description{
Filter a grid based on a list of restrictions
}
