% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{complete_temp_ts}
\alias{complete_temp_ts}
\title{Estimate the temperature on a given day for a given station}
\usage{
complete_temp_ts(temperature)
}
\arguments{
\item{temperature}{The output \code{tibble} of \code{process_temperature()}.}
}
\value{
The same \code{tibble} as the input but with completed temperature time series.
}
\description{
Estimate the temperature on a given day for a given station
}
\details{
The temperature time series are completed through the following approach:
\enumerate{
\item{Find a day with missing temperature in a time series of a given station.}
\item{Get the three closest stations that have non-missing temperature on the respective day.}
\item{Average the temperature of these three closest stations for the given day.}
}
}
