% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbour_matrices.R
\name{compute_nb_matrix}
\alias{compute_nb_matrix}
\title{Compute neighbourhood order matrices stratified by region and age group}
\usage{
compute_nb_matrix(regions, agegroups, geoms, nb_pattern, save_location = NULL)
}
\arguments{
\item{regions}{A \code{character} vector containing all region names.}

\item{agegroups}{A \code{character} vector containing all age group labels in increasing order.}

\item{geoms}{An object of class \code{sf} with at least a column named \code{region} that contains the unit names and a geometry column.}

\item{nb_pattern}{The neighbourhood \code{pattern} that is passed to \code{\link[sf]{st_relate}}.}

\item{save_location}{A filesystem path where the output matrix is saved.}
}
\value{
A matrix containing the neighbourhood order between the units
}
\description{
Compute neighbourhood order matrices stratified by region and age group
}
