% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{covariate_matrix}
\alias{covariate_matrix}
\title{Convert data tibble to matrix for use with Surveillance-package}
\usage{
covariate_matrix(data)
}
\arguments{
\item{data}{A \code{tibble} containing the columns \code{age}, \code{date}, \code{value} \code{region}.}
}
\value{
A matrix that can be used as a covariate in \code{hhh4} models or to fill the \code{observed} slot
in an \code{sts}-object.
}
\description{
Convert data tibble to matrix for use with Surveillance-package
}
