% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{df2matrix}
\alias{df2matrix}
\alias{matrix2df}
\title{Convert data tibble to matrix and back for use with \code{surveillance}-package}
\usage{
df2matrix(df)

matrix2df(mat, dates = NULL, reg_age = TRUE, format_orig = TRUE)
}
\arguments{
\item{df}{A \code{tibble} containing the columns \code{"age"}, \code{"date"}, \code{"value"} and \code{"region"}.}

\item{mat}{A \code{matrix} as for example the ones created by
\code{\link{matrix2df}}. Usually these are matrices as used for covariates in
\code{\link[surveillance]{hhh4}} models.}

\item{dates}{Either \code{NULL} (default) or a vector of class \code{Date}.
If dates are set in the rownames of \code{mat}, this argument is ignored. However,
if there are no date indications in the rownames of \code{mat}, this argument
must be provided. If rownames are valid dates, they are used and this argument
is ignored.}

\item{reg_age}{Must be either \code{TRUE} or \code{FALSE}.
If \code{TRUE} (default), unit names are interpreted as region and/or age
depending on whether unit names start with a digit (age) or a non-digit (region).
Unit names containing one or more dot(s) ('.') are interpreted as composed from different strata.
If there are different strata and \code{reg_age = TRUE}, the first two strata will be
interpreted as region and age, again depending on whether the colnames start with a
digit or a non-digit. If \code{FALSE}, stratum names are returned as "stratum_1", "stratum_2" etc.}

\item{format_orig}{This argument is only relevant if \code{reg_age = TRUE}.
Must be either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE}, the columns of
the returned tibble are brought into the same order as the ones from
the output of functions like for example \code{\link{get_cases}}.
This argument is specific for this master thesis and has no
practical use in other contexts.}
}
\value{
\code{df2matrix} returns a matrix that can be used as a covariate in \code{hhh4} models or to fill the \code{observed} slot
in an \code{sts}-object.

\code{matrix2df} returns a tibble with a columns "date", "value" and columns for each of th
strata used for the unit constructions.
}
\description{
Convert data tibble to matrix and back for use with \code{surveillance}-package
}
\examples{
urb <- get_urbanicity(time_res = "daily", spat_res = 1L, age_res = "age")
df2matrix(df = urb)
urb1 <- get_urbanicity("weekly", 1L, "no_age")
mat <- df2matrix(urb1)
urb2 <- matrix2df(mat)
all.equal(urb1, urb2)
}
