% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eurostat_query}
\alias{eurostat_query}
\title{Get data from Eurostat via the API instead of the bulk download facility}
\usage{
eurostat_query(dataset, args)
}
\arguments{
\item{dataset}{A \code{character} vector of length 1. The ID of a Eurostat data set.}

\item{args}{A \code{list} object. Contains further arguments used to filter \code{dataset}
server side.}
}
\value{
A \code{tibble} with the requested data.
}
\description{
Get data from Eurostat via the API instead of the bulk download facility
}
\details{
The argument \code{args} can be looked up using the
\href{https://ec.europa.eu/eurostat/web/json-and-unicode-web-services/getting-started/query-builder}{Eurostat query builder}.
}
\examples{
dataset <- "reg_area3"
args <- list(geoLevel = paste0("nuts", 1:3),
             filterNonGeo = "1",
             precision = "1",
             unit = "KM2",
             landuse = "TOTAL",
             shortlabel = "1",
             time = "2016")
area_size <- eurostat_query(dataset = dataset, args = args)

}
