% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_indicators.R
\name{get_boundary_inds}
\alias{get_boundary_inds}
\title{Make border indicators from geometries provided by Eurostat}
\usage{
get_boundary_inds(
  time_res = NULL,
  spat_res = NULL,
  age_res = NULL,
  cache_dir = NULL,
  enforce_cache = FALSE
)
}
\arguments{
\item{time_res}{A \code{character} vector of length 1 indicating the temporal resolution of the data.
Accepted values are either \code{"daily"} or \code{"weekly"}.}

\item{spat_res}{A \code{numeric} vector of length 1 indicating the spatial resolution of the data.
Accepted values are \code{0}, \code{1}, \code{2} and \code{3}. Corresponds to the respective NUTS level.}

\item{age_res}{A \code{character} vector of length 1 indicating whether or not the data should be stratified by
age groups. Accepted values are \code{"age"} and \code{"no_age"}.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{enforce_cache}{Either \code{TRUE} or \code{FALSE} (default). If \code{TRUE}, a cached version
of the data set is loaded. If there is no cached version, the function returns an error. If \code{FALSE},
the function returns a cached version if the data set is not older than a usual update cycle of the remote
source. Otherwise, the data is fetched from remote sources.}
}
\value{
A \code{list} of length 10. Each list element corresponds to a specific neighbour country or a general
spatial location at the country border. Each list element is a \code{tibble} with two columns \code{lvl3} and \code{value}. The
former indicates the respective NUTS 3 region whereas the latter indicates if the given region borders the specific country.
}
\description{
Make border indicators from geometries provided by Eurostat
}
\examples{
bounds <- get_boundary_inds()

}
