% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_indicators.R
\name{get_boundary_inds_from_source}
\alias{get_boundary_inds_from_source}
\title{Make border indicators from geometries provided by Eurostat}
\usage{
get_boundary_inds_from_source(cache_dir, filename)
}
\arguments{
\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{filename}{The name of the file that the border indicators are stored in.}
}
\value{
A \code{list} of length 10. Each list element corresponds to a specific neighbour country or a general
spatial location at the country border. Each list element is a \code{tibble} with two columns \code{lvl3} and \code{value}. The
former indicates the respective NUTS 3 region whereas the latter indicates if the given region borders the specific country.
}
\description{
Make border indicators from geometries provided by Eurostat
}
