% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_case_info}
\alias{get_case_info}
\title{Get the dates, age groups and regions present in the case data set}
\usage{
get_case_info(spat_res, time_res, cache_dir)
}
\arguments{
\item{spat_res}{The NUTS level on which to return the regions present in the case data set.}

\item{time_res}{The time resolution on which to return the dates present in the case data set.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("DEcovid_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("DEcovid_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(DEcovid_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{list} containing regions, age groups, and dates present in the case data set.
}
\description{
Get the dates, age groups and regions present in the case data set
}
