% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_colors.R
\name{get_map_colors}
\alias{get_map_colors}
\title{Color polygon geometries such that no element has the same color as all of its neighbours.}
\usage{
get_map_colors(geoms)
}
\arguments{
\item{geoms}{A single object of class \code{sf} or a list whose elements are of class \code{sf}. All of the \code{sf} objects must have two columns,
a geometry column and another column that is used as an Id for the objects. The Id column may be names or any other unique identifier.}
}
\value{
A \code{tibble} with columns \code{region} and \code{color} (contains color indices).
}
\description{
This function takes polygon geometries as input and returns a color index such that no two
neighbouring regions have the same color.
}
\details{
This function provides an efficient implementation of the DSATUR-algorithm developed by \insertCite{DEcovid:dsatur;textual}{DEcovid}.
}
\examples{
geoms <- get_geoms() # get geometries
geoms <- lapply(geoms, function(x) x[, c("region", "geometry")]) # remove the coloring
geoms <- get_map_colors(geoms)

}
\references{
\insertRef{DEcovid:dsatur}{DEcovid}
}
