% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_colors.R
\name{get_map_colors_from_source}
\alias{get_map_colors_from_source}
\title{Get map color indices}
\usage{
get_map_colors_from_source(geoms, nb_pattern, cache_dir, filename)
}
\arguments{
\item{geoms}{A single object of class \code{sf} or a list whose elements are of class \code{sf}.}

\item{nb_pattern}{A \code{character} vector of length one. Must contain a valid DE9-IM string describing the
relationship of \code{geoms}. This argument will be passed as the \code{pattern} argument to
\code{\link[sf:st_relate]{st_relate()}}.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{filename}{The name of the file where the color data is stored.}
}
\value{
A \code{tibble} with columns \code{region} and \code{color} (contains color indices).
}
\description{
Get map color indices
}
\details{
Under the hood, this function uses the function \code{getColoring}
from the package \href{https://github.com/hunzikp/MapColoring}{\code{MapColoring}}.
}
