% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbour_matrices.R
\name{get_nb_matrix}
\alias{get_nb_matrix}
\title{Compute neighbourhood order matrices stratified by region and/or age group}
\usage{
get_nb_matrix(by = c("age", "region"), nuts_level = 3, cache_dir = NULL)
}
\arguments{
\item{by}{A \code{character} vector. Contains one or both of \code{c("age", "region")}.}

\item{nuts_level}{A \code{character} vector of length 1 indicating the desired NUTS level of spatial regions. The argument is ignored if
\code{by = "age"}.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("DEcovid_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("DEcovid_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(DEcovid_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{matrix} containing the neighbourhood order of the units defined via the \code{by} argument.
}
\description{
Compute neighbourhood order matrices stratified by region and/or age group
}
\examples{
nb_regions_agegroups <- get_nb_matrix(nuts_level = 1)
nb_regions <- get_nb_matrix(by = "region", nuts_level = 1)
nb_age <- get_nb_matrix(by = "age", nuts_level = 1)

}
