% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urbanicity.R
\name{get_urbanicity_from_source}
\alias{get_urbanicity_from_source}
\title{Get urbanicity index from Destatis}
\usage{
get_urbanicity_from_source(cache_dir, filename)
}
\arguments{
\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{filename}{The name of the file where population data is stored.}
}
\value{
A \code{tibble} with columns \code{lvl3} and \code{value}. Column \code{lvl3} contains NUTS 3 districts,
whereas column \code{value} contains the log-population of the largest city within the respective district.
}
\description{
Get urbanicity index from Destatis
}
