% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{get_variants}
\alias{get_variants}
\title{Get a list with binary indicators of the dominant variant}
\usage{
get_variants(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{list} whose elements are time series indicating which variant was dominant in which week.
}
\description{
Get a list with binary indicators of the dominant variant
}
\examples{
variants <- get_variants()
}
