% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekdays.R
\name{get_weekend}
\alias{get_weekend}
\title{Get weekend covariate for a specific time period or the period relevant for this master thesis}
\usage{
get_weekend(
  spat_res = NULL,
  age_res = NULL,
  start = NULL,
  end = NULL,
  cache_dir = NULL
)
}
\arguments{
\item{spat_res}{A \code{numeric} vector of length 1 indicating the spatial resolution of the data.
Accepted values are \code{0}, \code{1}, \code{2} and \code{3}. Corresponds to the respective NUTS level.}

\item{age_res}{A \code{character} vector of length 1 indicating whether or not the data should be stratified by
age groups. Accepted values are \code{"age"} and \code{"no_age"}.}

\item{start}{The beginning of the time period. Must be formated as a "Date" vector of lenght 1.}

\item{end}{The end of the time period. Must be formated as a "Date" vector of lenght 1.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("DEcovid_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("DEcovid_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(DEcovid_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{tibble} with a date and a value column which contain the date sequence from start to end and a binary indicator
whether a day is a weekend day.
}
\description{
Get weekend covariate for a specific time period or the period relevant for this master thesis
}
\details{
All arguments default to NULL because the function has two use cases. Either the the arguments
\code{spat_res} and \code{age_res} have to be set or the arguments \code{start} and \code{end}.
}
\examples{
wke <- get_weekend(start = as.Date("2020-01-01"), end = as.Date("2020-12-31"))
wke <- get_weekend(age_res = "no_age", spat_res = 2)

}
