% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{grid2formulas}
\alias{grid2formulas}
\title{Convert a covariate grid of TRUEs and FALSEs into formulas}
\usage{
grid2formulas(grid, period = NULL)
}
\arguments{
\item{grid}{A \code{data.frame} with covariate names as column names and entries who take the values TRUE or FALSE.}

\item{period}{A vector of class \code{integer} and length 1. Must be positive and larger or equal to 1 if set.}
}
\value{
A \code{list} whose elements are formulas containing the covariates specified through the \code{grid}.
}
\description{
Convert a covariate grid of TRUEs and FALSEs into formulas
}
\details{
By default, the formulas always contain an intercept. However, adding random or unit-specific intercepts through the functions
\code{\link[surveillance]{ri}} or \code{\link[surveillance]{fe}} automatically removes the global intercept.
}
