% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\docType{data}
\name{holidays}
\alias{holidays}
\title{A data set containing binary indicators for public holidays for each day and NUTS 3 region in Germany}
\format{
A data frame with 451552 rows and 3 variables
}
\usage{
holidays
}
\description{
A data set containing binary indicators for public holidays for each day and NUTS 3 region in Germany
}
\keyword{datasets}
