% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{make_formulas}
\alias{make_formulas}
\title{Construct formulas for all \code{hhh4} model components}
\usage{
make_formulas(
  end = NULL,
  epi = NULL,
  ar = NULL,
  period = NULL,
  restrict = NULL
)
}
\arguments{
\item{end}{Either \code{NULL} or \code{character} vector listing all covariates that should be tried in the endemic component.}

\item{epi}{Either \code{NULL} or \code{character} vector listing all covariates that should be tried in the epidemic component.}

\item{ar}{Either \code{NULL} or \code{character} vector listing all covariates that should be tried in the autoregressive component.}

\item{period}{Either \code{NULL} or the period that is passed to \code{\link[surveillance]{addSeason2formula}}.}

\item{restrict}{Either \code{NULL} or an object of class \code{list} containing restrictions on the covariates passed in arguments \code{end}, \code{epi},
and \code{ar}. If not \code{NULL}, the \code{list} can have elements of  named "end", "epi" or "ar" which are again of class \code{list}
and determine restrictions to be applied to covariate combinations within each of the components. Restrictions must be stated
in the following form:
\describe{
\item{combined}{A \code{list} whose elements are \code{character} stating which covariates should be considered as one combined
covariate. This is useful in case of mutually exclusive indicators like weekday effects as it restricts the resulting formulas to
either contain all indicators or none of them.}
\item{always}{A \code{character} vector containing covariates that must always be included in the formula.}
\item{exclusive}{A \code{list} whose elements are again \code{lists} that contain mutually exclusive covariates. This
is useful as it allows the resulting formulas to contain maximally one of the covariates listed in these list elements.}
}

For further clarification on how to construct the restrictions list, see the examples section.}
}
\value{
A \code{data.frame} containing all possible combinations of formulas given model components and restrictions on
covariates.
}
\description{
Construct formulas for all \code{hhh4} model components
}
\examples{
weekdays <- paste0("wkd_", c("tu", "we", "th", "fr", "sa", "su"))
end <- c("vaccination", "temperature", "seasonality", "testing",
         "fe(1, unitSpecific = TRUE)", weekdays)
epi <- c("seasonality2", "testing", "stringency", "ri()", weekdays)
ar <- NULL
restrict <- list(end = list(combined = list(weekdays, c("testing", "vaccination")),
                            always = c("vaccination", weekdays),
                            exclusive = list(list(weekdays, "seasonality"),
                                             list("temperature", "seasonality"))),
                 epi = list(combined = list(weekdays),
                            always = c("seasonality2")))
formulas <- make_formulas(end = end, epi = epi, ar = ar,
                          period = 52, restrict = restrict)

}
