% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_path}
\alias{make_path}
\title{Construct path to a file from input directory and a file name}
\usage{
make_path(...)
}
\arguments{
\item{...}{The names of the directories and/or files in the correct order.}
}
\value{
Path to file for further use. Object of class \code{character}.
}
\description{
This function constructs valid paths from a
number of character strings that may or may not end with a directory separator.
}
