% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{needs_update}
\alias{needs_update}
\title{Check whether a file is older than a defined cutoff}
\usage{
needs_update(filepath, cutoff, units)
}
\arguments{
\item{filepath}{Object of class \code{character}. Indicates the path to the file to check.}

\item{cutoff}{Object of class \code{numeric}. How many days/hours/minutes etc. we accept the
file to be old.}

\item{units}{Object of class \code{character}. Unit of \code{cutoff}. One of the possibilities
listed in \code{\link[base:difftime]{base::difftime()}}.}
}
\value{
\code{TRUE} if the file is older than the cutoff and \code{FALSE} otherwise.
Object of class \code{logical}.
}
\description{
Returns \code{TRUE} if a file should be updated. Mainly to decide whether to
ignore cached versions of the file.
}
