% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\docType{data}
\name{overview}
\alias{overview}
\title{Overview over all datasets that the package \code{DEcovid} can download}
\format{
A data frame with 16 rows and 10 variables:
\describe{
\item{dataset}{The name of the dataset. However, this also includes unit specific intercepts as used in the author's master thesis.}
\item{time_sensitive}{A logical with values \code{TRUE} or \code{FALSE}. Indicates whether the values vary over time.}
\item{time_resolution}{A character vector with values \code{"daily"}, \code{"weekly"} or \code{NA}. Indicates the lowest possible time resolution at which the values
still vary over time. If the value for \code{time_sensitive} is \code{FALSE}, the corresponding value here is \code{NA}.}
\item{space_sensitive}{A logical with values \code{TRUE} or \code{FALSE}. Indicates whether the values vary over space.}
\item{space_resolution}{An integer value between 0 and 3. Indicates the lowest possible NUTS level at which the values
still vary over space. As NUTS 0 corresponds to the national level, a value of 0 indicates that the corresponding data does not vary over space at lower spatial
resolutions.}
\item{age_sensitive}{A logical with values \code{TRUE} or \code{FALSE}. Indicates whether the values vary over age groups.}
\item{age_resolution}{A character vector with values \code{"age"}, \code{"no_age"}. This column is used in some of the functions within the package. However,
the information is the same as in the column \code{age_sensitive}.}
\item{needs_fetching}{A logical with values \code{TRUE} or \code{FALSE}. Indicates whether the corresponding data set needs to be downloaded or whether it is implemented
in other packages implementing the Endemic-Epidemic model.}
\item{is_intercept}{A logical with values \code{TRUE} or \code{FALSE}. Primarily used for model fitting. Indicates whether the term in column \code{dataset} is
an intercept term or not.}
\item{include}{A \code{list} column that contains functions which take the arguments \code{spat_res}, \code{time_res} and \code{age_res} and return either \code{TRUE} or
\code{FALSE}. This is used to determine whether the corresponding data set can be included in an EE model with the corresponding resolution.}
}
}
\usage{
overview
}
\description{
This \code{tibble} contains an overview over all data sets that the package provides access to. It also
provides information about the resolutions that the data varies over for each of the spatial, temporal, and age group dimensions. It
also contains information about terms used for fitting Endemic-Epidemic (EE) models.
}
\keyword{datasets}
