% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{parameter_table}
\alias{parameter_table}
\title{Make a table from an \code{hhh4} model object}
\usage{
parameter_table(
  model,
  par_names,
  idx2Exp = FALSE,
  reparamPsi = TRUE,
  amplitudeShift = TRUE
)
}
\arguments{
\item{model}{\code{hhh4} model object}

\item{par_names}{A named \code{character} vector which is used to map variable names to the actual names for the table.}

\item{idx2Exp}{Must be one of \code{TRUE} or \code{FALSE}. Indicates whether covariates should be given on the exp scale.}

\item{reparamPsi}{Must be one of \code{TRUE} or \code{FALSE}. Indicates whether overdispersion parameters should be reparametrised.}

\item{amplitudeShift}{Must be one of \code{TRUE} or \code{FALSE}. Indicates whether seasonality coefficients should be reparametrised.}
}
\value{
A \code{tibble} containing parameter estimates and standard errors. Can be passed to \code{\link[kableExtra]{kable}}
}
\description{
Make a table from an \code{hhh4} model object
}
