% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot a fitted \code{hhh4} model}
\usage{
plot_fitted(x, units = "all", meanHHH = NULL)
}
\arguments{
\item{x}{A fitted \code{hhh4} model.}

\item{units}{Either \code{"all"}, \code{"aggregated"}, or a \code{character} vector indicating which units should be plotted.}

\item{meanHHH}{See \code{\link[surveillance]{plotHHH4_fitted}}.}
}
\value{
A ggplot object of the fitted model.
}
\description{
Plot a fitted \code{hhh4} model
}
