% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{process_temperature}
\alias{process_temperature}
\title{Import and process the data from the ECAD textfiles.}
\usage{
process_temperature(country, start_date, end_date, cache_dir, filename)
}
\arguments{
\item{country}{An ISO 3166-1 abbreviation for the country of interest.}

\item{start_date}{The start date of the period of interest formatted according to ISO 8601 given as a \code{character} vector of length 1.}

\item{end_date}{Either \code{NULL} (the default) or the end date of the period of interest formatted according to
ISO 8601 given as a \code{character} vector of length 1. If \code{NULL}, the function returns data for the period starting at \code{start_date} until
the latest point for which data is available.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("DEcovid_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("DEcovid_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(DEcovid_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{filename}{The filename under which the processed temperature data should be stored in \code{cache_dir}.}
}
\value{
The function returns a \code{tibble} containing weather stations in Germany that provide non-missing data
between the 1st of January and the end of the data set. All temperature time series are of the same length.
}
\description{
Import and process the data from the ECAD textfiles.
}
