% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_from_cache}
\alias{read_from_cache}
\title{Read file from cache directory or get data from source}
\usage{
read_from_cache(cache_dir, filename, cutoff, units)
}
\arguments{
\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{filename}{Object of class \code{character}. Name of the cached file.}

\item{cutoff}{Object of class \code{numeric}. How many days/hours/minutes etc. we accept the
file to be old.}

\item{units}{Object of class \code{character}. Unit of \code{cutoff}. One of the possibilities
listed in \code{\link[base:difftime]{difftime()}}.}
}
\value{
Object of class \code{logical}. Returns \code{TRUE} if the cached file exists and is
not older than \code{cutoff} in the specified \code{units}. Otherwise returns \code{FALSE}.
}
\description{
Returns a decision whether to get data from a cached file (\code{TRUE}) or
from source (\code{FALSE}). If \code{cache_dir} does not exist yet, the function also creates
the directory.
}
