% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rthesis.R
\docType{package}
\name{rthesis}
\alias{rthesis}
\title{Functionality used for the thesis of Felix Hofmann}
\description{
The package \pkg{rthesis} provides functions used in my thesis. Most of the functions are used to fetch data
from various sources. However, the package also contains functionality for processing data. Some functions
}
\section{Functions included}{

The most relevant functions are: \cr
\code{\link{get_cases}} \cr
\code{\link{nuts_table}} \cr
\code{\link{get_boundary_inds}} \cr
\code{\link{get_area_size}} \cr
\code{\link{get_geoms}} \cr
\code{\link{get_temperature}} \cr
\code{\link{get_holidays}} \cr
\code{\link{get_population}} \cr
\code{\link{get_stringency}} \cr
\code{\link{get_testing}} \cr
\code{\link{get_urbanicity}} \cr
\code{\link{get_vaccination}} \cr
\code{\link{get_map_colors}} \cr
}

