% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_functions.R
\name{summarise_data}
\alias{summarise_data}
\title{Aggregate data sets across time, space and age groups}
\usage{
summarise_data(data, time_res, spat_res, age_res, time_f, spat_f, age_f)
}
\arguments{
\item{data}{A \code{tibble} with columns \code{region}, \code{date}, \code{age} and \code{value}
which should be aggregated.}

\item{time_res}{The temporal resolution to which the data should be aggregated. Possible values are
\code{"daily"} and \code{"weekly"}.}

\item{spat_res}{The NUTS level to which the data should be aggregated. Possible values are
\code{0}, \code{1}, \code{2} and \code{3}. These numbers correspond to the NUTS level.}

\item{age_res}{Whether or not the data should be aggregated. Possible values are
\code{"age"} and \code{"no_age"}. The latter aggregates the data over age groups.}

\item{time_f}{A function used to aggregate along the temporal dimension.}

\item{spat_f}{A function used to aggregate along the spatial dimension.}

\item{age_f}{A function used to aggregate the age groups.}
}
\value{
A \code{tibble} with columns \code{region}, \code{date}, \code{age} and \code{value}.
}
\description{
Aggregate data sets across time, space and age groups
}
