% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{summarise_temp_to_region}
\alias{summarise_temp_to_region}
\title{Summarise the completed time series of all stations to NUTS-3 region specific temperature time series.}
\usage{
summarise_temp_to_region(temperature, tol, nb_pattern, cache_dir)
}
\arguments{
\item{temperature}{The output \code{tibble} of \code{complete_temp_ts()}.}

\item{tol}{The tolerance in kilometers from the NUTS-3 region border that is used to determine whether a weather station
is still considered as relevant for the temperature within a specific region.}

\item{nb_pattern}{The neighbourhood pattern that is used to calculate the neighbourhood structure.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{tibble} with columns \code{region} and \code{ts}. The column \code{ts} is a nested column and
contains the temperature time series for the respective NUTS-3 region.
}
\description{
Summarise the completed time series of all stations to NUTS-3 region specific temperature time series.
}
