% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{get_cases}
\alias{get_cases}
\title{Get the case data from RKI repository}
\usage{
get_cases(
  time_res = "daily",
  spat_res = 3,
  age_res = "age",
  cache_dir = NULL,
  enforce_cache = FALSE
)
}
\arguments{
\item{time_res}{A \code{character} vector of length 1 indicating the temporal resolution of the data.
Accepted values are either \code{"daily"} or \code{"weekly"}.}

\item{spat_res}{A \code{numeric} vector of length 1 indicating the spatial resolution of the data.
Accepted values are \code{0}, \code{1}, \code{2} and \code{3}. Corresponds to the respective NUTS level.}

\item{age_res}{A \code{character} vector of length 1 indicating whether or not the data should be stratified by
age groups. Accepted values are \code{"age"} and \code{"no_age"}.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{enforce_cache}{Either \code{TRUE} or \code{FALSE} (default). If \code{TRUE}, a cached version
of the data set is loaded. If there is no cached version, the function returns an error. If \code{FALSE},
the function returns a cached version if the data set is not older than a usual update cycle of the remote
source. Otherwise, the data is fetched from remote sources.}
}
\value{
A \code{tibble} with columns \code{age}, \code{date}, \code{value} (contains the counts) and \code{region}.
The variables are aggregated to the desired level.
}
\description{
Get the case data from RKI repository
}
\examples{
\dontrun{
cases <- get_cases(time_res = "daily", spat_res = 3, age_res = "age")
cases <- get_cases(time_res = "weekly", spat_res = 3, age_res = "age")
cases <- get_cases(time_res = "daily", spat_res = 1, age_res = "age")
cases <- get_cases(time_res = "daily", spat_res = 3, age_res = "no_age")
}


}
