% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{get_temperature}
\alias{get_temperature}
\title{Get temperature data set for this thesis}
\usage{
get_temperature(completed = FALSE, cache_dir = NULL)
}
\arguments{
\item{completed}{A \code{logical} of length 1 which indicates whether or not to estimate missing values from
neighbouring weather stations.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{tibble} containing all weather stations within Germany
with location and temperature time series starting on 1 January 2020.
}
\description{
Get temperature data set for this thesis
}
\examples{
\dontrun{
temperature <- get_temperature()
}
}
