% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{get_temperature}
\alias{get_temperature}
\title{Get temperature data set for this thesis}
\usage{
get_temperature(
  time_res = NULL,
  spat_res = NULL,
  age_res = NULL,
  complete = c("none", "station", "region"),
  tol = 0,
  cache_dir = NULL,
  enforce_cache = FALSE
)
}
\arguments{
\item{time_res}{A \code{character} vector of length 1 indicating the temporal resolution of the data.
Accepted values are either \code{"daily"} or \code{"weekly"}.}

\item{spat_res}{A \code{numeric} vector of length 1 indicating the spatial resolution of the data.
Accepted values are \code{0}, \code{1}, \code{2} and \code{3}. Corresponds to the respective NUTS level.}

\item{age_res}{A \code{character} vector of length 1 indicating whether or not the data should be stratified by
age groups. Accepted values are \code{"age"} and \code{"no_age"}.}

\item{complete}{A \code{character} vector of length 1 which determines the level of processing. If \code{"none"}, returns
the time series at each station with potential missing data. If \code{"station"}, the potentially missing temperatures are
completed by averaging the temperatures at the three closest weather stations that have non-missing data available for the
same day. If \code{"region"}, the temperature time series at all stations within a NUTS-3 region (plus those within \code{tol}
kilometers of the region boundary) are averaged in order to return a single temperature time series per NUTS-3 region. If there
is no station within a NUTS-3 region plus the tolerance, the time series for that region is calculated as an average of the
region's neighbours.}

\item{tol}{The tolerance in kilometers from the NUTS-3 region border that is used to determine whether a weather station
is still considered as relevant for the temperature within a specific region.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{enforce_cache}{Either \code{TRUE} or \code{FALSE} (default). If \code{TRUE}, a cached version
of the data set is loaded. If there is no cached version, the function returns an error. If \code{FALSE},
the function returns a cached version if the data set is not older than a usual update cycle of the remote
source. Otherwise, the data is fetched from remote sources.}
}
\value{
A \code{tibble} containing all weather stations within Germany
with location and temperature time series starting on 1 January 2020.
}
\description{
Get temperature data set for this thesis
}
\examples{
\dontrun{
temperature <- get_temperature(time_res = "weekly",
                               spat_res = 1L,
                               age_res = "no_age",
                               complete = "station")
}
}
