% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_path}
\alias{make_path}
\title{Construct path to a file from input directory and a file name}
\usage{
make_path(dir_path, filename)
}
\arguments{
\item{dir_path}{a string determining the path to the file directory.}

\item{filename}{a string determining the file name}
}
\value{
Path to file for further use. Object of class \code{character}.
}
\description{
This function constructs valid paths from a
path to a directory and the file name.
}
