% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{process_temperature}
\alias{process_temperature}
\title{Import and process the data from the ECAD textfiles.}
\usage{
process_temperature(cache_dir, filename)
}
\arguments{
\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("rthesis_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("rthesis_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(rthesis_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{filename}{The filename under which the processed temperature data should be stored in \code{cache_dir}.}
}
\value{
The function returns a \code{tibble} containing weather stations in Germany that provide non-missing data
between the 1st of January and the end of the data set. All temperature time series are of the same length.
}
\description{
Import and process the data from the ECAD textfiles.
}
